JWTを用いた認証

+++ hedgedoc-portal/src/main/java/com/atsushini/hedgedocportal/service/RuleService.java
@@ -4,12 +4,14 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.atsushini.hedgedocportal.dto.CurrentUserDto;
+import com.atsushini.hedgedocportal.dto.UserDto;
+import com.atsushini.hedgedocportal.authentication.AuthenticationUtil;
 import com.atsushini.hedgedocportal.dto.FolderDto;
 import com.atsushini.hedgedocportal.dto.RuleDto;
 import com.atsushini.hedgedocportal.entity.Folder;
 import com.atsushini.hedgedocportal.entity.Rule;
 import com.atsushini.hedgedocportal.entity.User;
+import com.atsushini.hedgedocportal.exception.ForbiddenException;
 import com.atsushini.hedgedocportal.exception.NotFoundException;
 import com.atsushini.hedgedocportal.repository.FolderRepository;
 import com.atsushini.hedgedocportal.repository.RuleRepository;
@@ -25,23 +27,29 @@ public class RuleService {
     private final FolderRepository folderRepository;
     private final UserRepository userRepository;
 
-    public List<RuleDto> getRules(CurrentUserDto userDto) {
-        List<Rule> ruleDtos = ruleRepository.findByUserId(userDto.getId());
+    public List<RuleDto> getRules() {
+        UserDto currentUserDto = AuthenticationUtil.getCurrentUser();
+        List<Rule> ruleDtos = ruleRepository.findByUserId(currentUserDto.getId());
 
         return ruleDtos.stream().map(this::convertToDto).toList();
     }
 
     // 振り分けルールを作成する
-    public void create(String title, String reqularExpression, Long folderId, CurrentUserDto currentUser) {
+    public void create(String title, String reqularExpression, Long folderId) {
         Folder folder = folderRepository.findById(folderId).orElse(null);
         if (folder == null) throw new NotFoundException("Folder not found with ID: " + folderId);        
+
+        UserDto currentUserDto = AuthenticationUtil.getCurrentUser();
+        if (!folder.getUser().getId().equals(currentUserDto.getId())) {
+            throw new ForbiddenException();
+        }
 
         Rule rule = new Rule();
         rule.setTitle(title);
         rule.setRegularExpression(reqularExpression);
         rule.setFolder(folder);
 
-        User user = userRepository.findById(currentUser.getId()).orElse(null);
+        User user = userRepository.findById(currentUserDto.getId()).orElse(null);
         rule.setUser(user);
 
         ruleRepository.save(rule);
@@ -49,11 +57,18 @@ public class RuleService {
 
     // 振り分けルールを更新する
     public void update(Long id, String title, String regularExpression, Long folderId) {
+        UserDto currenUserDto = AuthenticationUtil.getCurrentUser();
         Rule rule = ruleRepository.findById(id).orElse(null);
         if (rule == null) throw new NotFoundException("Rule not found with ID: " + id);
+        if (!rule.getUser().getId().equals(currenUserDto.getId())) {
+            throw new ForbiddenException();
+        }
 
         Folder folder = folderRepository.findById(folderId).orElse(null);
         if (folder == null) throw new NotFoundException("Folder not found with ID: " + folderId);        
+        if (!folder.getUser().getId().equals(currenUserDto.getId())) {
+            throw new ForbiddenException();
+        }
 
         rule.setTitle(title);
         rule.setRegularExpression(regularExpression);
@@ -66,6 +81,11 @@ public class RuleService {
     public void delete(Long id) {
         Rule rule = ruleRepository.findById(id).orElse(null);
         if (rule == null) throw new NotFoundException("Rule not found with ID: " + id);
+
+        UserDto currenUserDto = AuthenticationUtil.getCurrentUser();
+        if (!rule.getUser().getId().equals(currenUserDto.getId())) {
+            throw new ForbiddenException();
+        }
 
         ruleRepository.delete(rule);
     }
